The Coin Change Problem
### [Challenge Name: The Coin Change Problem](/challenges/coin-change)


You are working at the cash counter at a fun-fair, and you have different types of coins available to you in infinite quantities. The value of each coin is already given. Can you determine the number of ways of making change for a particular number of units using the given types of coins? 

For example, if you have $4$ types of coins, and the value of each type is given as $8, 3, 1, 2$ respectively, you can make change for $3$ units in three ways: 
$\{1, 1, 1\}$, $\{1, 2\}$, and $\{3\}$.

Complete the function `getWays` that takes the number of coin types and the value of each coin type as input, and return the number of ways to make change for $n$ units using any number of coins.